Prepare data for single RoC estimation

Create a single list with all information needed to estimate RoC. This is done because such list can be then evaluated in parallel.

Usage

prepare_data(
  data_source_prep,
  working_units = c("levels", "bins", "MW"),
  bin_size = 500,
  number_of_shifts = 5,
  rand = NULL
)

Arguments

data_source_prep

List with community and age

working_units

Character. Selection of units that the dissimilarity_coefficient will be calculated between.

"levels" - individual levels are going to be used

"bins" - samples in predefined bins will be pooled together and one sample will be selected from each time bin as a representation.

"MW" - Bins of selected size are created, starting from the beginning of the core. This is repeated many times, with each time bin (window) shifting by Z years forward. This is repeated X times, where X = bin size / Z.

bin_size

Numeric. Size of the time bin (in years)

number_of_shifts

Numeric. Value determining the number of shifts of window used in Moving window method

rand

Numeric. Number of runs used in randomisation.