Reduce datasets in merged format

Check the community dataset for all-zero taxa and empty levels and filter them out. This simplified version is optimised for performance.

Usage

reduce_data_simple(
  data_source_reduce,
  ommit_vars = c("label", "res_age", "age_diff"),
  check_taxa = TRUE,
  check_levels = TRUE
)

Arguments

data_source_reduce

list with community, age, and age_un

ommit_vars

character vector of column names to exclude from community data.

check_taxa

logical. Should columns be checked for all-zero taxa?

check_levels

logical. Should rows be checked for empty levels?

Value

The input data.frame with redundant columns and rows removed.