Run a single interation of RoC estimation

A single run is computed following the simple steps:

Subsetting levels in each bin: Here the working units (WU) are defined

Standardisation of assemblage data in each WU

Calculation of calculated as the dissimilarity coefficient (dissimilarity_coefficient)

Calculation of RoC between WUs: RoC is calculated as (dissimilarity_coefficient) standardised by age differences between WUs.

Usage

run_iteration(
  data_source_run,
  bin_selection = "first",
  standardise = FALSE,
  n_individuals = 150,
  tranform_to_proportions = TRUE,
  dissimilarity_coefficient = "euc",
  time_standardisation = 500,
  verbose = FALSE,
  silent = FALSE
)

Arguments

data_source_run

List with data and bins prepared by prepare_data

bin_selection

Character. Setting determining the the process of selection of samples from bins.

"first" - sample closest to the beginning of the bin is selected as a representation.

"random" - a random sample is selected as a representation.

standardise

Logical. If standardise == TRUE, then standardise each Working Unit to certain number of individuals (using random resampling without repetition)

n_individuals

Numeric. Number of grain to perform standardisation to. The N_individual is automatically adjusted to the smallest number of pollen grains in sequence.

tranform_to_proportions

Logical. Should the community data be transformed to a proportion during calculations?

dissimilarity_coefficient

Character. Dissimilarity coefficient. Type of calculation of differences between Working Units. See vegan::vegdist for more details.

"euc" - Euclidean distance

"euc.sd" - Standardised Euclidean distance

"chord" - Chord distance

"chisq" - Chi-squared coefficient

"gower" - Gower's distance

"bray" - Bray-Curtis distance

time_standardisation

Numeric. Units scaling for result RoC values. For example, if time_standardisation = 100, the RoC will be reported as dissimilarity per 100 yr.

verbose

Logical. If TRUE, function will output messages about internal processes

silent

Logical. If TRUE, suppress all console outputs (overrides verbose). Useful for testing.

See also