Smooth the community data

A function to apply one of the 4 smoothers.

Usage

smooth_community_data(
  data_source_smooth,
  smooth_method = c("m.avg", "grim", "age.w", "shep"),
  smooth_n_points = 5,
  smooth_n_max = 9,
  smooth_age_range = 500,
  round_results = FALSE,
  verbose = FALSE,
  silent = FALSE
)

Arguments

data_source_smooth

List with community, and age

smooth_method

Character. type of smoothing applied for the each of the pollen type

"none" - Pollen data is not smoothed

"m.avg" - Moving average

"grim" - Grimm's smoothing

"age.w"" - Age-weighted average

"shep" - Shepard's 5-term filter

smooth_n_points

Numeric. Number of points for used for moving average, Grimm and Age-Weighted smoothing (odd number)

smooth_n_max

Numeric. Maximal number of samples to look in Grimm smoothing

smooth_age_range

Numeric. Maximal age range for both Grimm and Age-weight smoothing

round_results

Logical. Should smoothed values be rounded to integers?

verbose

Logical. If TRUE, function will output messages about internal processes

silent

Logical. If TRUE, suppress all console outputs (overrides verbose). Useful for testing.

Details

Smoothing of assemblage data: Each variable within the assemblage data is smoothed using one of five in-built smoothing methods:

Shepard's 5-term filter (smooth_method = "shep"; Davis, 1986; Wilkinson, 2005)

moving average (smooth_method = "m.avg"})

age-weighted average (smooth_method = "age.w")

Grimm's smoothing (smooth_method = "grim"; Grimm and Jacobson, 1992)

References

Davis, J.C., 1986. Statistics and Data Analysis in Geology, 2nd edn. ed. J. Wiley & Sons, New York.

Grimm, E.C., Jacobson, G.L., 1992. Fossil-pollen evidence for abrupt climate changes during the past 18000 years in eastern North America. Clim. Dyn. 6, 179-184.

Wilkinson, L., 2005. The Grammar of Graphics. Springer-Verlag, New York, USA 37.

See also