Standardise the community data

Taxa in the assemblage dataset can be standardised to a certain count (e.g. number of pollen grains in each WU) by rarefaction. Random sampling without replacement is used to draw a selected number of individuals from each WU (e.g. 150 pollen grains).

Usage

standardise_community_data(data_source_standard, n_individuals = 150)

Arguments

data_source_standard

Data.frame with taxons as columns

n_individuals

Numeric. Number of grain to perform standardisation to. The N_individual is automatically adjusted to the smallest number of pollen grains in sequence.