RRatepol: Estimate rate of change

This function was deprecated, please use estimate_roc().

Usage

fc_estimate_RoC(
  data_source_community,
  data_source_age,
  age_uncertainty = NULL,
  smooth_method = c("none", "m.avg", "grim", "age.w", "shep"),
  smooth_n_points = 5,
  smooth_N_points = lifecycle::deprecated(),
  smooth_age_range = 500,
  smooth_n_max = 9,
  smooth_N_max = lifecycle::deprecated(),
  working_units = c("levels", "bins", "MW"),
  Working_Units = lifecycle::deprecated(),
  bin_size = 500,
  number_of_shifts = 5,
  Number_of_shifts = lifecycle::deprecated(),
  bin_selection = c("random", "first"),
  standardise = FALSE,
  n_individuals = 150,
  N_individuals = lifecycle::deprecated(),
  dissimilarity_coefficient = c("euc", "euc.sd", "chord", "chisq", "gower", "bray"),
  DC = lifecycle::deprecated(),
  tranform_to_proportions = TRUE,
  rand = NULL,
  use_parallel = FALSE,
  interest_threshold = NULL,
  time_standardisation = NULL,
  verbose = FALSE
)

Arguments

data_source_community

Data.frame. Community data with species as columns and levels (samples) as rows. First column should be sample_id (character).

data_source_age

Data.frame with two columns:

sample_id - unique ID of each level (character)

age - age of level (numeric)

age_uncertainty

Usage of age uncertainty form Age-depth models. Either:

matrix with number of columns as number of samples. Each column is one sample, each row is one age sequence from age-depth model. Age sequence is randomly sampled from age-depth model uncertainties at the beginning of each run.

NULL - Age uncertainties are not available and, therefore, will not be used.

smooth_method

Character. type of smoothing applied for the each of the pollen type

"none" - Pollen data is not smoothed

"m.avg" - Moving average

"grim" - Grimm's smoothing

"age.w"" - Age-weighted average

"shep" - Shepard's 5-term filter

smooth_n_points

Numeric. Number of points for used for moving average, Grimm and Age-Weighted smoothing (odd number)

smooth_N_points

smooth_N_points is no longer supported; please use smooth_n_points

smooth_age_range

Numeric. Maximal age range for both Grimm and Age-weight smoothing

smooth_n_max

Numeric. Maximal number of samples to look in Grimm smoothing

smooth_N_max

smooth_N_max is no longer supported; please use smooth_n_max

working_units

Character. Selection of units that the dissimilarity_coefficient will be calculated between.

"levels" - individual levels are going to be used

"bins" - samples in predefined bins will be pooled together and one sample will be selected from each time bin as a representation.

"MW" - Bins of selected size are created, starting from the beginning of the core. This is repeated many times, with each time bin (window) shifting by Z years forward. This is repeated X times, where X = bin size / Z.

Working_Units

Working_Units is no longer supported; please use working_units

bin_size

Numeric. Size of the time bin (in years)

number_of_shifts

Numeric. Value determining the number of shifts of window used in Moving window method

Number_of_shifts

Number_of_shifts is no longer supported; please use number_of_shifts

bin_selection

Character. Setting determining the the process of selection of samples from bins.

"first" - sample closest to the beginning of the bin is selected as a representation.

"random" - a random sample is selected as a representation.

standardise

Logical. If standardise == TRUE, then standardise each Working Unit to certain number of individuals (using random resampling without repetition)

n_individuals

Numeric. Number of grain to perform standardisation to. The N_individual is automatically adjusted to the smallest number of pollen grains in sequence.

N_individuals

N_individuals is no longer supported; please use n_individuals

dissimilarity_coefficient

Character. Dissimilarity coefficient. Type of calculation of differences between Working Units. See vegan::vegdist for more details.

"euc" - Euclidean distance

"euc.sd" - Standardised Euclidean distance

"chord" - Chord distance

"chisq" - Chi-squared coefficient

"gower" - Gower's distance

"bray" - Bray-Curtis distance

DC

DC is no longer supported; please use dissimilarity_coefficient

tranform_to_proportions

Logical. Should the community data be transformed to a proportion during calculations?

rand

Numeric. Number of runs used in randomisation.

use_parallel

Preference of usage of parallel computation of randomisation

[value] - selected number of cores

TRUE - automatically selected number of cores

FALSE - does not use parallel computation (only single core)

interest_threshold

Numeric. Optional. Age, after which all results of RoC are excluded.

time_standardisation

Numeric. Units scaling for result RoC values. For example, if time_standardisation = 100, the RoC will be reported as dissimilarity per 100 yr.

verbose

Logical. If TRUE, function will output messages about internal processes

See also